Improve CTCLoss + move loss and metrics to separate files
